user authorization

+++ app/Http/Controllers/ListingController.php
@@ -64,7 +64,9 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing) {
 
         // Make sure logged in user is owner
-        
+        if ($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized Action");
+        }
 
         $formFields = $request->validate([
             "title" => "required",
@@ -87,6 +89,12 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        
+         // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized Action");
+        } 
+
         $listing->delete();
         return redirect("/")->with("message", "Listing deleted successfully ");
     }
